UBERF-6598: Perform upgrade all workspaces to new versions
